% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.estimators.R
\name{calc.estimators}
\alias{calc.estimators}
\title{Calculating best K estimators for Evanno's and Pritchard's methods}
\usage{
calc.estimators(x)
}
\arguments{
\item{x}{a table returned from read.structure function}
}
\description{
this function grab data from read.structure
and returns a table with estimators calculated
}
