% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.best.to.plot.R
\name{find.best.to.plot}
\alias{find.best.to.plot}
\title{Finding the Best replicates to plot}
\usage{
find.best.to.plot(m, ...)
}
\arguments{
\item{m}{the output from the read.structure function}

\item{...}{parameters to be passed to nested functions}
}
\description{
screen the output of read.structure function and uses the lnPr of data to get the file name and plot the correspondent qmatrix
}
\details{
this function requires that the work directory would be the same of the structure output file.
}
