% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.modules.R
\name{find.modules}
\alias{find.modules}
\title{Find whether and how many modals are in each K-values}
\usage{
find.modules(path)
}
\arguments{
\item{path}{the path to structure results folder}
}
\description{
This function still prototype. find different modules for a given K value
}
