% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kview.R
\name{kview}
\alias{kview}
\title{Vizualize structure results}
\usage{
kview(dir, pop_names)
}
\arguments{
\item{dir}{a place or a zip file with structure results}

\item{pop_names}{a vector given the names of populations to be printed. It must be in same order and number as inputed populations in structure files.}
}
\description{
return a list with the main results of the pipeline.
}
