% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plot_estimators}
\alias{plot_estimators}
\title{Ploting Best-K Estimators}
\usage{
plot_estimators(m)
}
\arguments{
\item{m}{a table returned from calc.estimators function}
}
\description{
return a list with ggplot2 plots for all estimators
}
