% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.functions.R
\name{plotq}
\alias{plotq}
\title{Ploting the barplot coancestry cofficient}
\usage{
plotq(m = NULL, plot_colors = NULL, pop_names = NULL)
}
\arguments{
\item{m}{a q matrix extracted with qmatrix function}

\item{plot_colors}{a set of colors iqual to K to color each plot}

\item{pop_names}{a vector of population names to be ploted}
}
\description{
creates a plot of coancestry coficient with ggplot2
}
