% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmatrix.R
\name{qmatrix}
\alias{qmatrix}
\title{Extract Coancestry cofficient}
\usage{
qmatrix(file = NULL, ...)
}
\arguments{
\item{file}{a file path to one of the structure runs}

\item{...}{parameters to be used with other functions}
}
\description{
extract Q information for each K
}
