% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.structure.R
\name{read.structure}
\alias{read.structure}
\title{Reading structure output}
\usage{
read.structure(x)
}
\arguments{
\item{x}{a directory path or a zip file containing the result log
files from structure runs}
}
\description{
reads files from a structure run and returns a summary table
}
\details{
this is the most primary basic function.
}
